pyproject.toml
[project]
name = "sia-tp2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.4",
    "pillow>=12.2.0",
]

[dependency-groups]
dev = [
    "mypy>=1.20.0",
]
